# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'The PracticeProject API'
  description: 'This API was initially developed for an entrance test and is now being used for practice. The skills practiced are: unit tests, documentation generation and JSON schemas.'
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
  /api/auth/login:
    post:
      summary: 'Send a login request.'
      operationId: sendALoginRequest
      description: "This endpoint allows you to try login.\nUnauthenticated users can't do anything, except try login or register."
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: 'valid credentials'
          content:
            text/plain:
              schema:
                type: string
                example: '{ [ "token": "[a_JWT_Token]" ] }'
        400:
          description: 'invalid credentials'
          content:
            text/plain:
              schema:
                type: string
                example: '{ [ ] }'
      tags:
        - 'Auth management'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'The email of the user.'
                  example: example@example.com
                password:
                  type: string
                  description: 'The password of the user.'
                  example: pa$$word
              required:
                - email
                - password
      security: []
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  -
    default: []